Update Securing-Domain-Controllers-Against-Attack.md
We need to drop the section on security configuration wizard. it was deprecated over ten years ago.
Adding long version of abbreviation RODC.
We also need to remove references to old OS versions and provide some links.

Wow, a number of old and even some stale links...

Note more review may be required to replace SCW to configure the firewall (deletion of second to last paragraph).